create sub-files for functions related to BoardValue
